+++ src/options/common.rs
@@ -3,6 +3,7 @@ mod cmd;
 mod logging;
 mod network;
 mod search;
+mod time;
 mod value;
 
 pub use address::*;
@@ -10,4 +11,5 @@ pub use cmd::*;
 pub use logging::*;
 pub use network::*;
 pub use search::*;
+pub use time::*;
 pub use value::*;