+++ src/options/config/client/api.rs
@@ -1,6 +1,8 @@
 use serde::{Deserialize, Serialize};
 
+use crate::options::common::Seconds;
+
 #[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
 pub struct ClientApiConfig {
-    pub timeout: Option<f32>,
+    pub timeout: Option<Seconds>,
 }