+++ src/options/config/server.rs
@@ -3,7 +3,10 @@ use serde::{Deserialize, Serialize};
 use super::common::LoggingSettings;
 
 mod listen;
+mod watch;
+
 pub use listen::*;
+pub use watch::*;
 
 /// Represents configuration settings for the distant server
 #[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
@@ -12,4 +15,5 @@ pub struct ServerConfig {
     pub logging: LoggingSettings,
 
     pub listen: ServerListenConfig,
+    pub watch: ServerWatchConfig,
 }